Delete extra default language relationships created by PLL_CRUD_Terms::save_term() hook
When new tags and categories are imported they are assigned the default
language - PLL_CRUD_Terms::save_term() hook takes care of that.  If a
non-default language is set in an imported 'term_translations' later the
superfluous default needs to be removed. Otherwise the term would be linked to
TWO languages simultaneously.