Delete extra translation groups creted by PLL_CRUD_Terms::save_term() hook
When new tags and categories are imported they are assigned the default
language (previous commit fixes it) and a new translation group is created for
each of them which consists of just one entry - the new tag/category being
created.  Similarily to the default language when the proper translation group
is later post-processed in PLL_WP_Import::remap_translations() the
default-created single-entry groups need to be removed. They are not only no
longer needed but also break admin UI which doesn't display the names of
translated entries when editing the newly imported tag/category (probably break
language switchers too).